requirements.txt
asgiref==3.5.2
Authlib==0.14.1
boto3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==3.3.2
debugpy
Django==4.1
django-environ==0.9.0
django-storages
drf-yasg
djangorestframework>=3.13.1,!=3.14.0
idna==2.9
pillow
pre-commit
psycopg2-binary>=2.8
pycparser==2.20
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.10
PyJWT==2.1.0
pyjwt[crypto]
cryptography
